authenticaion and authorization

+++ backend/controller/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 
 // @desc GET goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
 
     res.status(200).json(goals);
 });
@@ -22,7 +23,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     });
 
     res.status(201).json(goal);
@@ -39,6 +41,18 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error("Goal not found");
     }
 
+    // check for user
+    if (!req.user) {
+        res.user(401);
+        throw new Error('User not found');
+    }
+
+    // make sure logged in user matches the goal user
+    if (goal.user.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true });
 
     res.status(200).json(updatedGoal);
@@ -53,6 +67,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400);
         throw new Error("Goal not found");
+    }
+
+    // check for user
+    if (!req.user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+
+    // make sure the logged in user matches the goal user
+    if (goal.user.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
     }
 
     await goal.deleteOne();
